requirements.txt
# Core CLI and UI libraries
click>=8.1.0
rich>=13.0.0
inquirer>=3.1.0
pyyaml>=6.0

# Additional utilities
requests>=2.31.0
gitpython>=3.1.0
pathlib2>=2.3.0
